changed getMovies into hook
